+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+<<<<<<< HEAD
 ## [0.13.0](https://github.com/access-ci-org/access-ci-account/compare/v0.12.0...v0.13.0) (2026-07-30)
 
 
@@ -8,12 +9,23 @@
 * **auth:** return user to previous URL after login ([e64a533](https://github.com/access-ci-org/access-ci-account/commit/e64a533f524737ae7d5fae8691b40ebdb6d3dc8b))
 
 ## [0.12.0](https://github.com/access-ci-org/access-ci-account/compare/v0.11.0...v0.12.0) (2026-07-29)
+=======
+## [0.14.0-beta1](https://github.com/access-ci-org/access-ci-account/compare/v0.11.0...v0.14.0-beta1) (2026-07-30)
+>>>>>>> ramps-1120-recovery-email-updated
 
 
 ### Features
 
 * update My ACCESS menu ([df7e143](https://github.com/access-ci-org/access-ci-account/commit/df7e1435034f63d1b27be54e5c1d728e69e4754d))
 
+<<<<<<< HEAD
+=======
+
+### Miscellaneous Chores
+
+* release 0.14.0-beta1 ([a0beb57](https://github.com/access-ci-org/access-ci-account/commit/a0beb572d445603877394cff1d3de3e062cf2c52))
+
+>>>>>>> ramps-1120-recovery-email-updated
 ## [0.11.0](https://github.com/access-ci-org/access-ci-account/compare/v0.10.2...v0.11.0) (2026-07-29)